+++ Tests/DataPipeline/DataPipelineCompatibilityTests.swift
@@ -1,4 +1,5 @@
 @testable import CioAnalytics
+@testable import CioInternalCommonMocks
 @testable import CioDataPipelines
 @testable import CioInternalCommon
 import Foundation
@@ -500,7 +501,9 @@ private extension DataPipelineCompatibilityTests {
 
 private extension SavedEvent {
     var eventType: String? { self[keyPath: "type"] as? String }
-    subscript(mapKeyPath keyPath: KeyPath) -> [String: Any]? { value(keyPath: keyPath, reference: nil) as? [String: Any] }
+    subscript(mapKeyPath keyPath: KeyPath) -> [String: Any]? {
+        value(keyPath: keyPath, reference: nil) as? [String: Any]
+    }
 
     /// checks recursively if a given key exists anywhere in the event
     func containsKey(_ key: String) -> Bool {